# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cambridge AI Tutor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscription/select-plan:
    post:
      operationId: selectSubscriptionPlan
      tags: [subscription]
      summary: Switch the current student onto a different plan
      description: >-
        Activates immediately. No payment is collected yet — see
        SubscriptionPlan.priceCents.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectPlanBody"
      responses:
        "200":
          description: Updated subscription usage on the newly selected plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionUsage"
        "400":
          description: Unknown plan id
  /subscription/checkout:
    post:
      operationId: createSubscriptionCheckout
      tags: [subscription]
      summary: Start a Stripe Checkout session to upgrade to a paid plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionBody"
      responses:
        "200":
          description: Checkout URL to redirect the student to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Unknown or free plan id
        "503":
          description: Stripe is not configured
  /subscription/portal:
    get:
      operationId: getSubscriptionPortal
      tags: [subscription]
      summary: Get a link to the Stripe-hosted billing portal (manage card, invoices, cancel)
      responses:
        "200":
          description: Portal URL to redirect the student to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Student has no Stripe customer yet (never checked out)
        "503":
          description: Stripe is not configured
...
  /subscription/topup/checkout:
    post:
      operationId: createTopupCheckout
      tags: [subscription]
      summary: Start a Stripe Checkout session to purchase a credit pack
      description: Only allowed when the student's current plan has allowsTopups=true.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupCheckoutBody"
      responses:
        "200":
          description: Checkout URL to redirect the student to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Unknown credit pack id
        "403":
          description: Current plan does not allow purchasing top-ups
        "503":
          description: Stripe is not configured
...
components:
...
  schemas:
...
    SubscriptionFeature:
      type: string
      description: A quota-metered AI feature.
      enum: [tutorMessage, noteGeneration, quizGeneration, flashcardGeneration, oralPractice]
    SubscriptionFeatureUsage:
      type: object
      properties:
        feature:
          $ref: "#/components/schemas/SubscriptionFeature"
        used:
          type: integer
          description: Units consumed so far in the current billing period.
        limit:
          type: integer
          description: Units included in the student's plan for this feature, per billing period.
        remaining:
          type: integer
          description: Units left before the student hits their limit for this feature.
      required: [feature, used, limit, remaining]
    SubscriptionUsage:
      type: object
      properties:
        planId:
          type: string
        planName:
          type: string
        status:
          type: string
          enum: [active, canceled, past_due, expired]
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
        allowsTopups:
          type: boolean
          description: Whether the student's current plan permits purchasing extra credit packs.
        usage:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionFeatureUsage"
      required:
        - planId
        - planName
        - status
        - currentPeriodStart
        - currentPeriodEnd
        - allowsTopups
        - usage
...
    CreateCheckoutSessionBody:
      type: object
      properties:
        planId:
          type: string
      required: [planId]
    TopupCheckoutBody:
      type: object
      properties:
        creditPackId:
          type: integer
      required: [creditPackId]
    CheckoutSessionResponse:
      type: object
      properties:
        url:
          type: string
          format: uri
      required: [url]
...
    SelectPlanBody:
      type: object
      properties:
        planId:
          type: string
      required: [planId]